+++ scylla/src/cluster/mod.rs
@@ -1,1 +1,5 @@
+mod cluster;
 pub mod metadata;
+
+pub(crate) use cluster::{Cluster, ClusterNeatDebug};
+pub use cluster::{ClusterData, Datacenter};
